feat: ímplemented lb3 (changed the repository dependency injection implementation to the Entity Framework).

+++ web-app-csharp/Entities/DepartmentEntity.cs
@@ -2,7 +2,7 @@ namespace web_app_csharp.Entities;
 
 public class DepartmentEntity
 {
-    public int DeptId { get; set; }
+    public decimal DeptId { get; set; }
 
     public string Name { get; set; }
